requirements.txt
kafka-python-ng
pyspark
python-dotenv